refactoring tracker into own package
- created tracker package
- moved Tracker files to new Tracker Package
- updated imports and paths for Trackers in User package
